Refactor: 난수 생성과 SHA-256 해싱 기능을 추가한 CryptoUtils 구현
